Update WebHelper.cpp
For easy identification of an aircraft in the network it is good to use specific known aircraft addresses (FLARM or ICAO ones), rather than a randomly generated ones. The firmware is modified so that the user can enter the address of the aircraft and the type of the address. Alternatively, a randomly denerated address still can be used.